+++ app/src/main/java/org/horaapps/leafpic/util/Security.java
@@ -1,17 +1,14 @@
 package org.horaapps.leafpic.util;
 
 import android.content.Context;
-import android.content.DialogInterface;
 import android.graphics.PorterDuff;
-import android.os.Build;
+import android.os.CancellationSignal;
 import android.os.IBinder;
-import android.support.annotation.RequiresApi;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.CardView;
 import android.view.View;
 import android.view.inputmethod.InputMethodManager;
 import android.widget.EditText;
-import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import com.orhanobut.hawk.Hawk;
@@ -19,6 +16,7 @@ import com.orhanobut.hawk.Hawk;
 import org.horaapps.leafpic.R;
 import org.horaapps.liz.ThemeHelper;
 import org.horaapps.liz.ThemedActivity;
+import org.horaapps.liz.ui.ThemedIcon;
 
 import java.security.MessageDigest;
 
@@ -27,59 +25,56 @@ import java.security.MessageDigest;
  */
 public class Security {
 
-    public static void setPasswordOnDelete(Context context, boolean passwordOnDelete) {
+    public static void setPasswordOnDelete(boolean passwordOnDelete) {
         Hawk.put("password_on_delete", passwordOnDelete);
     }
 
-    public static void setPasswordOnHidden(Context context, boolean passwordOnHidden) {
+    public static void setPasswordOnHidden(boolean passwordOnHidden) {
         Hawk.put("password_on_hidden", passwordOnHidden);
     }
 
-    public static void setFingerprintUnlock(Context context, boolean passwordOnHidden) {
+    public static void setFingerprintUnlock(boolean passwordOnHidden) {
         Hawk.put("fingerprint_security", passwordOnHidden);
     }
 
-    public static boolean isPasswordSet(Context context) {
+    public static boolean isPasswordSet() {
         return Hawk.get("password_hash", null) != null;
     }
 
-    public static boolean isPasswordOnHidden(Context context) {
+    public static boolean isPasswordOnHidden() {
         return Hawk.get("password_hash", null) != null && Hawk.get("password_on_hidden", false);
     }
 
-    public static boolean isFingerprintUsed(Context context) {
+    public static boolean isFingerprintUsed() {
         return Hawk.get("fingerprint_security", false);
     }
 
-    public static boolean isPasswordOnDelete(Context context) {
+    public static boolean isPasswordOnDelete() {
         return Hawk.get("password_hash", null) != null && Hawk.get("password_on_delete", false);
     }
 
-    private static boolean checkPassword(Context context, String pass) {
+    private static boolean checkPassword(String pass) {
         return sha256(pass).equals(Hawk.get("password_hash", null));
     }
 
-    public static boolean setPassword(Context context, String newValue) {
+    public static boolean setPassword(String newValue) {
         return Hawk.put("password_hash", sha256(newValue));
     }
 
-    public static boolean clearPassword(Context context) {
+    public static boolean clearPassword() {
         return Hawk.delete("password_hash");
     }
 
-    public static void askPassword(final ThemedActivity activity, final PasswordInterface passwordInterface) {
-        AlertDialog.Builder passwordDialogBuilder = new AlertDialog.Builder(activity, activity.getDialogStyle());
+    public static void authenticateUser(final ThemedActivity activity, final AuthCallBack passwordInterface) {
+        AlertDialog.Builder builder = new AlertDialog.Builder(activity, activity.getDialogStyle());
+        CancellationSignal mCancellationSignal = new CancellationSignal();
 
-        if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.M) {
-            FingerprintHandler fingerprintHandler = new FingerprintHandler(activity);
-        } else {
+        final View view = activity.getLayoutInflater().inflate(org.horaapps.leafpic.R.layout.dialog_password, null);
+        final TextView passwordDialogTitle = view.findViewById(org.horaapps.leafpic.R.id.password_dialog_title);
+        final CardView passwordDialogCard = view.findViewById(org.horaapps.leafpic.R.id.password_dialog_card);
+        final EditText editTextPassword = view.findViewById(org.horaapps.leafpic.R.id.password_edittxt);
+        final ThemedIcon fingerprintIcon = view.findViewById(R.id.fingerprint_icon);
 
-        }
-
-        final View PasswordDialogLayout = activity.getLayoutInflater().inflate(org.horaapps.leafpic.R.layout.dialog_password, null);
-        final TextView passwordDialogTitle = (TextView) PasswordDialogLayout.findViewById(org.horaapps.leafpic.R.id.password_dialog_title);
-        final CardView passwordDialogCard = (CardView) PasswordDialogLayout.findViewById(org.horaapps.leafpic.R.id.password_dialog_card);
-        final EditText editTextPassword = (EditText) PasswordDialogLayout.findViewById(org.horaapps.leafpic.R.id.password_edittxt);
 
         passwordDialogTitle.setBackgroundColor(activity.getPrimaryColor());
         passwordDialogCard.setBackgroundColor(activity.getCardBackgroundColor());
@@ -88,106 +83,56 @@ public class Security {
         editTextPassword.setTextColor(activity.getTextColor());
 
 
-        passwordDialogBuilder.setView(PasswordDialogLayout);
+        builder.setView(view);
 
-        passwordDialogBuilder.setPositiveButton(activity.getString(R.string.ok_action).toUpperCase(), new DialogInterface.OnClickListener() {
-            public void onClick(DialogInterface dialog, int which) {
-                // NOTE: set this empty, later will be overwrite to avoid the dismiss
-            }
+        builder.setPositiveButton(activity.getString(R.string.ok_action).toUpperCase(), (dialog, which) -> {
+            // NOTE: set this empty, later will be overwrite to avoid the dismiss
         });
 
-        passwordDialogBuilder.setNegativeButton(activity.getString(R.string.cancel).toUpperCase(), new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                hideKeyboard(activity, editTextPassword.getWindowToken());
-            }
-        });
+        builder.setNegativeButton(activity.getString(R.string.cancel).toUpperCase(), (dialog, which) -> hideKeyboard(activity, editTextPassword.getWindowToken()));
 
-        final AlertDialog passwordDialog = passwordDialogBuilder.create();
-        passwordDialog.show();
+        final AlertDialog dialog = builder.create();
+        dialog.show();
         showKeyboard(activity);
         editTextPassword.requestFocus();
 
-        passwordDialog.getButton(AlertDialog.BUTTON_POSITIVE).setOnClickListener(new View
-                .OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (checkPassword(activity, editTextPassword.getText().toString())) {
-                    hideKeyboard(activity, editTextPassword.getWindowToken());
-                    passwordDialog.dismiss();
-                    passwordInterface.onSuccess();
-                } else {
-                    editTextPassword.getText().clear();
-                    editTextPassword.requestFocus();
-                    passwordInterface.onError();
-                }
+        if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.M && Security.isFingerprintUsed()) {
+            FingerprintHandler fingerprintHandler = new FingerprintHandler(activity, mCancellationSignal);
+            if (fingerprintHandler.isFingerprintSupported()) {
+                fingerprintHandler.setOnFingerprintResult(new FingerprintHandler.CallBack() {
+                    @Override
+                    public void onSuccess() {
+                        dialog.dismiss();
+                        passwordInterface.onAuthenticated();
+                    }
+
+                    @Override
+                    public void onError(String s) {
+                        // TODO: 9/9/17 siplaymessage
+                    }
+                });
+
+                fingerprintHandler.startAuth();
+            } else {
+                fingerprintIcon.setVisibility(View.GONE);
+            }
+        } else {
+            fingerprintIcon.setVisibility(View.GONE);
+        }
+
+        dialog.getButton(AlertDialog.BUTTON_POSITIVE).setOnClickListener(v -> {
+            if (checkPassword(editTextPassword.getText().toString())) {
+                hideKeyboard(activity, editTextPassword.getWindowToken());
+                mCancellationSignal.cancel();
+                dialog.dismiss();
+                passwordInterface.onAuthenticated();
+            } else {
+                editTextPassword.getText().clear();
+                editTextPassword.requestFocus();
+                passwordInterface.onError();
             }
         });
     }
-
-    @RequiresApi(api = Build.VERSION_CODES.M)
-    public class AskFingerprint implements FingerprintHandler.OnFingerprintResult {
-
-        AlertDialog passwordDialog;
-        FingerprintInterface fingerprintInterface;
-
-        public AskFingerprint() {
-        }
-
-        public void ask(final ThemedActivity activity, final FingerprintInterface fingerprintInterface) {
-            AlertDialog.Builder passwordDialogBuilder = new AlertDialog.Builder(activity, activity.getDialogStyle());
-
-            this.fingerprintInterface = fingerprintInterface;
-            FingerprintHandler fingerprintHandler = new FingerprintHandler(activity);
-            fingerprintHandler.setOnFingerprintResult(this);
-            fingerprintHandler.startAuth();
-
-            final View PasswordDialogLayout = activity.getLayoutInflater().inflate(R.layout.dialog_fingerprint, null);
-            final TextView passwordDialogTitle = (TextView) PasswordDialogLayout.findViewById(org.horaapps.leafpic.R.id.password_dialog_title);
-            final CardView passwordDialogCard = (CardView) PasswordDialogLayout.findViewById(org.horaapps.leafpic.R.id.password_dialog_card);
-            final LinearLayout llUsePsw = (LinearLayout) PasswordDialogLayout.findViewById(R.id.ll_use_psw);
-            final ThemedIcon fingerprintIcon = (ThemedIcon) PasswordDialogLayout.findViewById(R.id.fingerprint_icon);
-
-            fingerprintIcon.setColor(activity.getIconColor());
-
-            passwordDialogTitle.setBackgroundColor(activity.getPrimaryColor());
-            passwordDialogCard.setBackgroundColor(activity.getCardBackgroundColor());
-
-            passwordDialogBuilder.setView(PasswordDialogLayout);
-
-
-            passwordDialogBuilder.setNegativeButton(activity.getString(R.string.cancel).toUpperCase(), new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(DialogInterface dialog, int which) {
-                    // NOTE: set this empty, later will be overwrite to avoid the dismiss
-                }
-            });
-
-
-            passwordDialog = passwordDialogBuilder.create();
-            passwordDialog.show();
-
-            llUsePsw.setOnClickListener(view -> {
-                passwordDialog.dismiss();
-                fingerprintInterface.onUsePsw();
-            });
-
-        }
-
-        @Override
-        public void onSuccess() {
-            passwordDialog.dismiss();
-            fingerprintInterface.onSuccess();
-        }
-
-        @Override
-        public void onError() {
-            fingerprintInterface.onError();
-
-        }
-    }
-
-
     private static void showKeyboard(Context context) {
         InputMethodManager imm = (InputMethodManager) context.getSystemService(Context.INPUT_METHOD_SERVICE);
         imm.toggleSoftInput(InputMethodManager.SHOW_FORCED, 0);
@@ -215,17 +160,8 @@ public class Security {
     }
 
 
-    public interface PasswordInterface {
-        void onSuccess();
-
+    public interface AuthCallBack {
+        void onAuthenticated();
         void onError();
-    }
-
-    public interface FingerprintInterface {
-        void onSuccess();
-
-        void onError();
-
-        void onUsePsw();
     }
 }